+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Added validation for approved content pack usecases and tags.
+* Added new code validations for *CommonServerPython* import to `XSOAR-linter`.
 
 
 # v1.2.9